+++ yggdrasil_decision_forests/port/python/ydf/model/model.cc
@@ -155,7 +155,8 @@ void init_model(py::module_& m) {
       // WARNING: This method releases the Global Interpreter Lock.
       .def("Benchmark", WithStatusOr(&GenericCCModel::Benchmark),
            py::arg("dataset"), py::arg("benchmark_duration"),
-           py::arg("warmup_duration"), py::arg("batch_size"))
+           py::arg("warmup_duration"), py::arg("batch_size"),
+           py::arg("num_threads"))
       .def("VariableImportances", &GenericCCModel::VariableImportances)
       .def("ForceEngine", &GenericCCModel::ForceEngine, py::arg("engine_name"))
       .def("ListCompatibleEngines", &GenericCCModel::ListCompatibleEngines)